CocoaLumberjack setup with colored output support
See https://github.com/CocoaLumberjack/CocoaLumberjack

Colored output is awesome, I can't work without it.

I rekon is important to set up the logging framework as the first step of the development because otherwise the cost of doing it increases with the number of NSLogs that are gonna be written in the meantime.