Loading more posts automatically when the user reaches the bottom
Because I already extracted the delegate, and the UITableViewDelegate is the UIScrollViewDelegate as well, I had two choices: either bring back the delegate in the view controller, or find a way to run view controller code from the delegate.

I really didn't want to bring the delegate back in, it would have made the view controller code less clean. So I decided to use a block implementation. This keeps the delegate ignorant of what the view controller does or is, although the code now is a tiny bit more _convoluted_.